Split JavaScript runtime launch unwrapper
